Expire login item appropriately
When fetching balances, make sure to mark the login item as expired in case the link is expired.
It is possible that we miss the webhook update from plaid. Since we are fetching the balances
anyways, we should then make sure to properly account for link expirations as a redundency
mechanism.